cambios generales

+++ app/Http/Controllers/TarifaController.php
@@ -5,6 +5,11 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Yajra\Datatables\Datatables;
 use App\Models\Tarifa;
+use App\Models\Pasarela;
+use App\Models\Aplicacion;
+use App\Models\Dispositivo;
+use App\Models\Log;
+use App\Models\Servicio;
 
 class TarifaController extends Controller
 {
@@ -43,6 +48,7 @@ class TarifaController extends Controller
         if ($user->can('Actualizar tarifas')) {
             $tarifa = Tarifa::find($id);
             $tarifa->update($request->all());
+            Log::saveLogs('Tarifas', 'Actualizar', $tarifa->id);
             return back()->with('success', 'Tarifa actualizada correctamente');
         }
         abort(403, 'No tienes permiso para ver esta página');
@@ -53,6 +59,11 @@ class TarifaController extends Controller
         $user = auth()->user();
         if ($user->can('Eliminar tarifas')) {
             $tarifa = Tarifa::find($id);
+            $tarifas = Servicio::where('id_tarifa', $id)->get();
+            if (count($tarifas) > 0) {
+                return back()->with('error', 'No se puede eliminar la tarifa porque está asociada a un servicio');
+            }
+            Log::saveLogs('Tarifas', 'Eliminar', $tarifa->id);
             $tarifa->delete();
             return back()->with('success', 'Tarifa eliminada correctamente');
         }
@@ -64,6 +75,7 @@ class TarifaController extends Controller
         $user = auth()->user();
         if ($user->can('Crear tarifas')) {
             $tarifa = Tarifa::create($request->all());
+            Log::saveLogs('Tarifas', 'Crear', $tarifa->id);
             return back()->with('success', 'Tarifa creada correctamente');
         }
         abort(403, 'No tienes permiso para ver esta página');
@@ -73,6 +85,27 @@ class TarifaController extends Controller
     {
         $tarifas = Tarifa::get();
 
-        return  response()->json($tarifas);
+        return response()->json($tarifas);
+    }
+
+    public function obtenerPasarelasJSON()
+    {
+        $pasarelas = Pasarela::get();
+
+        return response()->json($pasarelas);
+    }
+
+    public function obtenerAplicacionesJSON()
+    {
+        $aplicaciones = Aplicacion::get();
+
+        return response()->json($aplicaciones);
+    }
+
+    public function obtenerDispositivosJSON()
+    {
+        $dispositivos = Dispositivo::get();
+
+        return response()->json($dispositivos);
     }
 }
